Fix imports in documentation to include /ngx
The imports in the documentation were wrong, causing some people to create issues. I ran into this issue myself and did not know what was going wrong. The imports should have /ngx on the end in the readme, as this is what works with the latest version of the plugin.